Change namespace because of conflicting class
