+++ .rubocop.yml
@@ -36,3 +36,24 @@ Style/SymbolArray:
 
 Style/WordArray:
   EnforcedStyle: brackets
+
+########### METRICS ###############
+
+Metrics/AbcSize:
+  Max: 50
+
+Metrics/ClassLength:
+  Enabled: false
+
+Metrics/CyclomaticComplexity:
+  Max: 15
+
+Metrics/PerceivedComplexity:
+  Max: 15
+
+Metrics/MethodLength:
+  CountAsOne:
+    - array
+    - hash
+    - heredoc
+  Max: 30
