+++ lib/storage_tables/attachable/changes.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module StorageTables
+  module Attachable
+    module Changes # :nodoc:
+      extend ActiveSupport::Autoload
+
+      eager_autoload do
+        autoload :CreateOne
+      end
+    end
+  end
+end
